Crate bf

bf-rs is a optimizing Brainfuck interpreter and JIT compiler inspired by Eli Bendersky’s series on JIT compilation. It includes a library crate bf that exports most of the functionality, and an executable bfi that provides a command-line interface for executing Brainfuck programs.

By default, installing bf does not enable the JIT compiler because that requires nightly Rust. To build and install from crates.io with the JIT enabled:

$ rustup run nightly cargo install --features=jit bf

This library implements a number of compilation passes:

First, Brainfuck concrete syntax is parsed into an abstract syntax tree.

Then, repeated sequences of the same command are run-length encoded.

Then, common loop forms are converted to new (non-Brainfuck) instructions by the peephole optimizer.

The peephole output can be flattened to bytecode, which is then interpreted.

Or, if the jit feature is enabled (nightly only), the peephole output can be just-in-time compiled to x64 machine code.

Interpreters are provided for the intermediate forms as well. In particular, all representations of Brainfuck programs implement the Interpretable trait.

Modules

ast
Parsing and interpretation for unoptimized Brainfuck abstract syntax trees.

common
Definitions common to multiple passes.

flat
Brainfuck bytecode, flat instead of an abstract syntax tree.

jit
Just-in-time compiles Brainfuck AST to x64 machine code (--features jit, nightly only)

peephole

rle

state

test_helpers

traits