Type Definition bf::peephole::Program

type Program = [Statement];

At this level, a program is a rose tree of statements.

All instructions are leaves except for the Loop instruction, which contains a boxed Program.