Struct bf::state::State

pub struct State { /* fields omitted */ }

The BF machine state.

Methods

impl State

fn new() -> Self

Creates a new BF machine state with capacity [DEFAULT_CAPACITY].

fn with_capacity(capacity: usize) -> Self

Creates a new BF machine state.

fn left(&mut self, count: usize) -> BfResult<()>

Decrements the pointer.

Errors

Return Err if pointer would go below 0.

fn right(&mut self, count: usize) -> BfResult<()>

Increments the pointer.

Errors

Return Err if pointer would go past the end of the memory.

fn up(&mut self, count: u8)

Increments the byte at the pointer.

Wraps around modulo 256.

fn down(&mut self, count: u8)

Decrements the byte at the pointer.

Wraps around modulo 256.

fn load(&self) -> u8

Gets the value of the point at the pointer.

fn store(&mut self, value: u8)

Sets the value of the byte at the pointer.

fn up_pos_offset(&mut self, offset: usize, value: u8) -> BfResult<()>

Adds the given value at the given positive offset from the pointer.

fn up_neg_offset(&mut self, offset: usize, value: u8) -> BfResult<()>

Adds the given value at the given negative offset from the pointer.

fn read<R: Read>(&mut self, input: &mut R)

Reads from a Read into the byte at the pointer.

fn write<W: Write>(&self, output: &mut W)

Writes to a Write from the byte at the pointer.

fn capacity(&self) -> usize

The memory capacity.

fn as_mut_ptr(&mut self) -> *mut u8

Gets a mutable, raw pointer to the start of memory.