Struct bf::state::State

pub struct State { /* fields omitted */ }

The Brainfuck machine state.

Methods

impl State

pub fn new() -> Self

Creates a new BF machine state with memory capacity DEFAULT_CAPACITY (30_000).

pub fn with_capacity(memory_size: usize) -> Self

Creates a new BF machine state with the given memory capacity.

pub fn left<C: IntoUsize>(&mut self, count: C) -> BfResult<()>

Decrements/decreases the pointer.

Errors

Return Err if pointer would go below 0.

pub fn right<C: IntoUsize>(&mut self, count: C) -> BfResult<()>

Increments/increases the pointer.

Errors

Return Err if pointer would go past the end of the memory.

pub fn up(&mut self, count: u8)

Increments/increases the byte at the pointer.

Wraps around modulo 256.

pub fn down(&mut self, count: u8)

Decrements/decreases the byte at the pointer.

Wraps around modulo 256.

pub fn load(&self) -> u8

Gets the value of the byte at the pointer.

pub fn store(&mut self, value: u8)

Sets the value of the byte at the pointer.

pub fn up_pos_offset<C: IntoUsize>(
&mut self,
offset: C,
value: u8
) -> BfResult<()>

Adds the given value at the given positive offset from the pointer.

pub fn up_neg_offset<C: IntoUsize>(
&mut self,
offset: C,
value: u8
) -> BfResult<()>

Adds the given value at the given negative offset from the pointer.

pub fn read<R: Read>(&mut self, input: &mut R)

Reads from a Read into the byte at the pointer.

pub fn write<W: Write>(&self, output: &mut W)

Writes to a Write from the byte at the pointer.

pub fn capacity(&self) -> usize

The memory capacity.

pub fn as_mut_ptr(&mut self) -> *mut u8

Gets a mutable, raw pointer to the start of memory.

This is used by the JIT RTS to pass the memory pointer to the generated code.